Relocate dynamic form fields to own component
